--- a/fe/fe-core/src/test/java/com/starrocks/transaction/ConcurrentTxnNewPublishTest.java
+++ b/fe/fe-core/src/test/java/com/starrocks/transaction/ConcurrentTxnNewPublishTest.java
@@ -1,0 +1,10 @@
+package com.starrocks.transaction;
+
+import java.sql.SQLException;
+
+public class ConcurrentTxnNewPublishTest extends ConcurrentTxnTest {
+    @Override
+    boolean getEnableNewPublish() throws SQLException {
+        return true;
+    }
+}