--- a/fe/fe-core/src/test/java/com/starrocks/transaction/NewPublishTest.java
+++ b/fe/fe-core/src/test/java/com/starrocks/transaction/NewPublishTest.java
@@ -17,7 +17,6 @@ public class NewPublishTest {
     public static void setUp() throws Exception {
         int fePort = new Random().nextInt(10000) + 50000;
         PseudoCluster.getOrCreate("pseudo_cluster", fePort, 3);
-        Config.enable_new_publish_mechanism = true;
     }
 
     @AfterClass
@@ -27,6 +26,7 @@ public class NewPublishTest {
 
     @Test
     public void testInsertUsingNewPublish() throws Exception {
+        Config.enable_new_publish_mechanism = true;
         Connection connection = PseudoCluster.getInstance().getQueryConnection();
         Statement stmt = connection.createStatement();
         try {
@@ -34,8 +34,14 @@ public class NewPublishTest {
             stmt.execute("use test");
             stmt.execute(
                     "create table test ( pk bigint NOT NULL, v0 string not null, v1 int not null ) primary KEY (pk) DISTRIBUTED BY HASH(pk) BUCKETS 3 PROPERTIES(\"replication_num\" = \"3\", \"storage_medium\" = \"SSD\");");
-            Assert.assertFalse(stmt.execute("insert into test values (1,\"1\", 1), (2,\"2\",2), (3,\"3\",3);"));
-            System.out.printf("updated %d rows\n", stmt.getUpdateCount());
+            int numLoad = 100;
+            long startTs = System.nanoTime();
+            for (int i = 0; i < numLoad; i++) {
+                Assert.assertFalse(stmt.execute("insert into test values (1,\"1\", 1), (2,\"2\",2), (3,\"3\",3);"));
+            }
+            double t = (System.nanoTime() - startTs) / 1e9;
+            System.out.printf("numLoad:%d Time: %.2fs, %.2f tps\n", numLoad, t, numLoad / t);
+
         } finally {
             stmt.close();
             connection.close();