--- a/fe/pom.xml
+++ b/fe/pom.xml
@@ -732,6 +732,13 @@ under the License.
                 <version>5.1.38</version>
             </dependency>
 
+            <!-- https://mvnrepository.com/artifact/org.apache.commons/commons-dbcp2 -->
+            <dependency>
+                <groupId>org.apache.commons</groupId>
+                <artifactId>commons-dbcp2</artifactId>
+                <version>2.9.0</version>
+            </dependency>
+
         </dependencies>
     </dependencyManagement>