## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] refactor
- [ ] others

## Problem Summary(Required) ：
This PR adds some UT for concurrent load, it creates a database with 100/1000 tables and routinely loads each table using `insert into`, to compare txn handling performance, and persistent log size.
Note, this is not a real-world test, FE and BE are pseudo FE & BE, so it only partially reflects some performance implications of some FE classes.

```
100 tables
old publish:
numLoad:627 Time: 10.01s, 62.65 tps
Avg TXN state log length: 1483

new publish:
numLoad:627 Time: 10.01s, 62.66 tps
Avg TXN state log length: 1524
```

```
1000 tables
old publish:
numLoad:5732 Time: 17.22s, 332.79 tps

new publish:
numLoad:5732 Time: 17.56s, 326.39 tps

```
Using BDBJE rather than fake journal
```
1000 tables

old publish:
numLoad:5732 Time: 57.79s, 99.19 tps

new publish:
numLoad:5732 Time: 56.39s, 101.66 tps

```

```
add some BEBJE related metrics, number is median(if not specified) 

1DB
numLoad:5732 Time: 60.07s, 95.42 tps
numLog:18265 writeBatch:1.000000 writeLatency:2.000000 writeBytes:323.000000

20DB
numLoad:5732 Time: 42.55s, 134.71 tps
numLog:17220 writeBatch:1.000000 writeLatency:2.000000 writeBytes:440.000000

```