@decster 2022-07-29
Also clearly, there are some bottlenecks in TXN commit/publish code path, which limits total TPS. 
Currently, 300 TPS is enough for most use cases, optimizing TPS is not the priority.