@decster 2022-07-29
run starrocks_admit_test