@decster 2022-08-01
run starrocks_admit_test