@nshangyiming 2022-08-01
As showed in the test results, there is almost no tps increase with new publish mechanism, is this normal?