@decster 2022-08-01
> As showed in the test results, there is almost no tps increase with new publish mechanism, is this normal?

The purpose of the new publish is not TPS, and currently, the bottleneck is not clear, maybe DB lock or journal log